Split progress code
