+++ rhubarb/src/audio/processing.h
@@ -3,7 +3,7 @@
 #include <vector>
 #include <functional>
 #include "AudioClip.h"
-#include "tools/ProgressBar.h"
+#include "tools/progress.h"
 
 void process16bitAudioClip(const AudioClip& audioClip, std::function<void(const std::vector<int16_t>&)> processBuffer, size_t bufferCapacity, ProgressSink& progressSink);
 void process16bitAudioClip(const AudioClip& audioClip, std::function<void(const std::vector<int16_t>&)> processBuffer, ProgressSink& progressSink);
