+++ rhubarb/src/audio/voiceActivityDetection.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "AudioClip.h"
 #include "time/BoundedTimeline.h"
-#include "tools/ProgressBar.h"
+#include "tools/progress.h"
 
 JoiningBoundedTimeline<void> detectVoiceActivity(const AudioClip& audioClip, int maxThreadCount, ProgressSink& progressSink);
