+++ rhubarb/src/tools/ProgressBar.h
@@ -2,45 +2,8 @@
 
 #include <atomic>
 #include <future>
-#include <functional>
-#include <list>
-#include <vector>
-#include <mutex>
-#include <string>
 #include <iostream>
-
-class ProgressSink {
-public:
-	virtual ~ProgressSink() {}
-	virtual void reportProgress(double value) = 0;
-};
-
-class NullProgressSink : public ProgressSink {
-public:
-	void reportProgress(double) override {}
-};
-
-class ProgressForwarder : public ProgressSink {
-public:
-	ProgressForwarder(std::function<void(double progress)> callback);
-	void reportProgress(double value) override;
-private:
-	std::function<void(double progress)> callback;
-};
-
-class ProgressMerger {
-public:
-	ProgressMerger(ProgressSink& sink);
-	ProgressSink& addSink(double weight);
-private:
-	void report();
-
-	ProgressSink& sink;
-	std::mutex mutex;
-	double totalWeight = 0;
-	std::list<ProgressForwarder> forwarders;
-	std::vector<double> weightedValues;
-};
+#include "progress.h"
 
 class ProgressBar : public ProgressSink {
 public:
